Add files via upload
A non-metasploit exploit for the ms08-067-netapi vulnerability. Feel free to swap or omit the handler, it was just because I don't like using multiple windows for a single exploit. This exploit opens an administrator bind shell on port 443, plaintext. Feel free to swap the shellcode to - just read the details in the comments well before doing so. 

This exploit can exploit more than 2 platforms, it's just I didn't want to add the others. Read the source if you need to know. 

Note that not all of this is my code - this is just my modifications for an existing exploit. The original can be found at https://www.exploit-db.com/exploits/40279/ - my modifications are really just swapped shellcode, an included netcat bind shell handler, and some other minor tweaks. 

DO NOT USE FOR MALICIOUS PURPOSES. I do not condone illegal usage of this code, nor any other code on my page. I distribute this under the assumption of the purpose of either study or operation in a legal pentest or laboratory environment.